backend/requirements.txt
#### `backend/requirements.txt`
```
Flask==3.0.0
mysql-connector-python==8.2.0